webapp/api/requirements.txt
bottle
pymysql
requests
pyjwt
phonenumbers
uuid
